Giang

+++ src/main/java/com/shop/medicineshop/exception/ProductNotFoundException.java
@@ -1,6 +1,6 @@
 package com.shop.medicineshop.exception;
 public class ProductNotFoundException extends RuntimeException{
-    public ProductNotFoundException(Long id){
+    public ProductNotFoundException(Integer id){
         super("Could not found the user with id "+ id);
     }
 }
